+++ libraries/WaveMix/WaveMix.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: WaveMix.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 // PURPOSE: Arduino library to mix two signals (A and B) with an adaptive weight.
 //     URL: https://github.com/RobTillaart/WaveMix
 //
@@ -14,6 +14,8 @@ WaveMix::WaveMix()
 {
   _weight[0] = 0.5;
   _weight[1] = 0.5;
+  _gain      = 1.0;
+  _offset    = 0.0;
 }
 
 
@@ -34,10 +36,13 @@ void WaveMix::setPercentage(float percentage)
 
 float WaveMix::mix(float s1, float s2)
 {
+  float total = 0;
   if (_gain == 0) return 0;
-  if (_weight[0] == 0) return s2 * _gain;
-  if (_weight[1] == 0) return s1 * _gain;
-  return ((_weight[0] * s1) + (_weight[1] * s2)) * _gain;
+  if (_weight[0] != 0) total += (_weight[0] * s1);
+  if (_weight[1] != 0) total += (_weight[1] * s2);
+  total *= _gain;
+  if (_offset != 0) total += _offset;
+  return total;
 }