+++ libraries/WaveMix/WaveMix.h
@@ -2,7 +2,7 @@
 //
 //    FILE: WaveMix.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.1.1
+// VERSION: 0.1.2
 // PURPOSE: Arduino library to mix two signals (A and B) with an adaptive weight.
 //     URL: https://github.com/RobTillaart/WaveMix
 //
@@ -10,7 +10,7 @@
 
 #include "Arduino.h"
 
-#define WAVEMIX_LIB_VERSION               (F("0.1.1"))
+#define WAVEMIX_LIB_VERSION               (F("0.1.2"))
 
 
 class WaveMix
@@ -28,12 +28,16 @@ public:
   void  setGain(float gain) { _gain = gain; };
   float getGain() { return _gain; };
 
+  void  setOffset(float offset) { _offset = offset; };
+  float getOffset() { return _offset; };
+
   float mix(float s1, float s2 = 0);
 
 
 private:
   float _weight[2] = { 0.5, 0.5 };
   float _gain      = 1.0;
+  float _offset    = 0.0;
 
 };